+++ src/MIABIS_on_FHIR/diagnosis_report.py
@@ -6,11 +6,10 @@ from fhirclient.models.diagnosticreport import DiagnosticReport
 from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.meta import Meta
 
-from src.MIABIS_on_FHIR.util._constants import DEFINITION_BASE_URL
-from src.MIABIS_on_FHIR.util._parsing_util import get_nested_value, parse_reference_id
-from src.MIABIS_on_FHIR.util._util import create_fhir_identifier
 from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_reference_id
+from src.MIABIS_on_FHIR.util.util import create_fhir_identifier
 
 
 class DiagnosisReport:
