+++ src/MIABIS_on_FHIR/network.py
@@ -5,10 +5,10 @@ from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.group import Group
 from fhirclient.models.meta import Meta
 
-from src.MIABIS_on_FHIR.util._parsing_util import get_nested_value, parse_reference_id
-from src.MIABIS_on_FHIR.util._util import create_fhir_identifier
 from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_reference_id
+from src.MIABIS_on_FHIR.util.util import create_fhir_identifier
 
 
 class Network:
