+++ src/MIABIS_on_FHIR/sample.py
@@ -9,13 +9,13 @@ from fhirclient.models.fhirreference import FHIRReference
 from fhirclient.models.meta import Meta
 from fhirclient.models.specimen import Specimen, SpecimenCollection, SpecimenProcessing
 
-from src.MIABIS_on_FHIR.util._constants import DETAILED_MATERIAL_TYPE_CODES, DEFINITION_BASE_URL
-from src.MIABIS_on_FHIR.util._parsing_util import get_nested_value, parse_reference_id
-from src.MIABIS_on_FHIR.util._util import create_fhir_identifier, create_codeable_concept, \
-    create_codeable_concept_extension
 from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
 from src.MIABIS_on_FHIR.storage_temperature import StorageTemperature
-from src.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.constants import DETAILED_MATERIAL_TYPE_CODES
+from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value, parse_reference_id
+from src.MIABIS_on_FHIR.util.util import create_fhir_identifier, create_codeable_concept, \
+    create_codeable_concept_extension
 
 
 class Sample:
