+++ src/MIABIS_on_FHIR/sample_donor.py
@@ -6,12 +6,12 @@ from fhirclient.models.fhirdate import FHIRDate
 from fhirclient.models.meta import Meta
 from fhirclient.models.patient import Patient
 
-from src.MIABIS_on_FHIR.util._constants import DONOR_DATASET_TYPE, DEFINITION_BASE_URL
-from src.MIABIS_on_FHIR.util._parsing_util import get_nested_value
-from src.MIABIS_on_FHIR.util._util import create_fhir_identifier, create_codeable_concept_extension
 from src.MIABIS_on_FHIR.gender import Gender
 from src.MIABIS_on_FHIR.incorrect_json_format import IncorrectJsonFormatException
-from src.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.config import FHIRConfig
+from src.MIABIS_on_FHIR.util.constants import DONOR_DATASET_TYPE
+from src.MIABIS_on_FHIR.util.parsing_util import get_nested_value
+from src.MIABIS_on_FHIR.util.util import create_fhir_identifier, create_codeable_concept_extension
 
 
 class SampleDonor:
