+++ src/blaze_client/NonExistentResourceException.py
