+++ src/blaze_client/__init__.py
@@ -1,0 +1,2 @@
+from .blaze_client import BlazeClient
+from .NonExistentResourceException import NonExistentResourceException
